+++ showcase/src/main/resources/META-INF/externs/showcase.externs.js
@@ -66,6 +66,30 @@ Component.prototype.illustration
 /** @type {string} */
 Component.prototype.summary
 
+// ------------------------------------------------------ coordinates
+
+/** @constructor */
+function Coordinates() {
+}
+
+/** @type {string} */
+Coordinates.prototype.latitude
+
+/** @type {string} */
+Coordinates.prototype.longitude
+
+// ------------------------------------------------------ date of birth
+
+/** @constructor */
+function DateOfBirth() {
+}
+
+/** @type {string} */
+DateOfBirth.prototype.date
+
+/** @type {number} */
+DateOfBirth.prototype.age
+
 // ------------------------------------------------------ extension
 
 /** @constructor */
@@ -87,6 +111,30 @@ Extension.prototype.clazz
 /** @type {string} */
 Extension.prototype.summary
 
+// ------------------------------------------------------ file
+
+/** @constructor */
+function File() {
+}
+
+/** @type {string} */
+File.prototype.id
+
+/** @type {string} */
+File.prototype.name
+
+/** @type {string} */
+File.prototype.description
+
+/** @type {string} */
+File.prototype.icon
+
+/** @type {!Object<string, string>} */
+File.prototype.meta
+
+/** @type {!Array<!File>} */
+File.prototype.children
+
 // ------------------------------------------------------ layout
 
 /** @constructor */
@@ -107,3 +155,213 @@ Layout.prototype.clazz
 
 /** @type {string} */
 Layout.prototype.summary
+
+// ------------------------------------------------------ location
+
+/** @constructor */
+function Location() {
+}
+
+/** @type {!Street} */
+Location.prototype.street
+
+/** @type {string} */
+Location.prototype.city
+
+/** @type {string} */
+Location.prototype.state
+
+/** @type {string} */
+Location.prototype.postcode
+
+/** @type {!Coordinates} */
+Location.prototype.coordinates
+
+/** @type {!Timezone} */
+Location.prototype.timezone
+
+// ------------------------------------------------------ login
+
+/** @constructor */
+function Login() {
+}
+
+/** @type {string} */
+Login.prototype.uuid
+
+/** @type {string} */
+Login.prototype.username
+
+/** @type {string} */
+Login.prototype.password
+
+/** @type {string} */
+Login.prototype.salt
+
+/** @type {string} */
+Login.prototype.md5
+
+/** @type {string} */
+Login.prototype.sha1
+
+/** @type {string} */
+Login.prototype.sha256
+
+// ------------------------------------------------------ name
+
+/** @constructor */
+function Name() {
+}
+
+/** @type {string} */
+Name.prototype.title
+
+/** @type {string} */
+Name.prototype.first
+
+/** @type {string} */
+Name.prototype.last
+
+// ------------------------------------------------------ picture
+
+/** @constructor */
+function Picture() {
+}
+
+/** @type {string} */
+Picture.prototype.large
+
+/** @type {string} */
+Picture.prototype.medium
+
+/** @type {string} */
+Picture.prototype.thumbnail
+
+// ------------------------------------------------------ record
+
+/** @constructor */
+function Record() {
+}
+
+/** @type {string} */
+Record.prototype.title
+
+/** @type {string} */
+Record.prototype.released
+
+/** @type {number} */
+Record.prototype.year
+
+/** @type {string} */
+Record.prototype.url
+
+/** @type {string} */
+Record.prototype.cover
+
+/** @type {!Array<!Track>} */
+Record.prototype.tracks
+
+// ------------------------------------------------------ server
+
+/** @constructor */
+function Server() {
+}
+
+/** @type {string} */
+Server.prototype.name
+
+/** @type {number} */
+Server.prototype.threads
+
+/** @type {number} */
+Server.prototype.applications
+
+/** @type {number} */
+Server.prototype.workspaces
+
+/** @type {string} */
+Server.prototype.status
+
+/** @type {string} */
+Server.prototype.location
+
+// ------------------------------------------------------ street
+
+/** @constructor */
+function Street() {
+}
+
+/** @type {number} */
+Street.prototype.number
+
+/** @type {string} */
+Street.prototype.name
+
+// ------------------------------------------------------ timezone
+
+/** @constructor */
+function Timezone() {
+}
+
+/** @type {string} */
+Timezone.prototype.offset
+
+/** @type {string} */
+Timezone.prototype.description
+
+// ------------------------------------------------------ track
+
+/** @constructor */
+function Track() {
+}
+
+/** @type {number} */
+Track.prototype.track
+
+/** @type {string} */
+Track.prototype.title
+
+/** @type {!Array<string>} */
+Track.prototype.writer
+
+/** @type {string} */
+Track.prototype.length
+
+// ------------------------------------------------------ user
+
+/** @constructor */
+function User() {
+}
+
+/** @type {string} */
+User.prototype.gender
+
+/** @type {!Name} */
+User.prototype.name
+
+/** @type {!Location} */
+User.prototype.location
+
+/** @type {string} */
+User.prototype.email
+
+/** @type {!Login} */
+User.prototype.login
+
+/** @type {!DateOfBirth} */
+User.prototype.dob
+
+/** @type {!DateOfBirth} */
+User.prototype.registered
+
+/** @type {string} */
+User.prototype.phone
+
+/** @type {string} */
+User.prototype.cell
+
+/** @type {!Picture} */
+User.prototype.picture
+
+/** @type {string} */
+User.prototype.nat